requirements.txt
django~=3.1
gunicorn
django-heroku
requests~=2.24.0
djangorestframework~=3.11.1
